Add experimental Cohere support for sessions + prefix option
Turns out that Cohere's model being a third the size of OpenAI's makes
it much more difficult to get good results for the default AI chat
agent. Support for it is being left in because a lot of the
restructuring of the code will make it easier to add other models as
they (if they?) get released.

The prefix_user/prefix_ai options were also added to try and assist with
making Coheres model a bit more usable.