Add serialization of the initial CLI arguments to the session file.
Prior to this change the session file contained a serialized copy of the
default CompletionOptions struct that you could change to override the
default unspecified options. We've changed things to serialize the
initial CLI arguments to the session file to remember them. This helps
keep continuity between sessions for options like "temperature" and
"system"