Add support for calling other prompts asynchronously.
Adding support for this broke the underlying ai libraries assumption that
their was only one chat running between the user and server. Now multiple
different prompt files can be running independently to process input
from other prompts, and sending responses to eachother, which are then
processed by the AI agent server. To tackle this confusion, the watch
module was added that will monitor the underlying session files to pick
up on writes from other prompts running at the same time, so we can get
a running log of all the input happening on one prompt. It automatically
loads up just the main prompt (the first declared prompt) for now, but
support should be broadened in the future to listen to the other prompts
declared in a prompt file.

The animal house example was temporarily simplified in order to test two
prompts just echoing back and forth at eachother, and it seems to work
fairly well so far.

The api_request example was also modified to make it work a little more
consistently, but it's still not 100%.