Fix bug with YAML options parsing.
If you started your options with a wrapped multiline string like:
```
bob
    description: >
        This didn't work
```
The parse would have the incorrect indentation for the first line and
then be off when starting the multiline string, which YAML does not
like, so we removed whitespace parsing from after prompt names in order
to get the correct indentation.

This somehow broke putting comments directly after prompt name's. For
now I do not care, no one's going to do that anyway it looks ugly as
hell.